add base view mvc
