+++ app/src/main/java/com/example/pengalatdite/aplikasisederhana/ItemViewMvc.java
@@ -2,7 +2,7 @@ package com.example.pengalatdite.aplikasisederhana;
 
 import android.view.View;
 
-public interface ItemViewMvc {
+public interface ItemViewMvc extends ViewMvc {
     View getRootView();
 
     void bindData(DataNews dataNews, String position);
