+++ app/src/main/java/com/example/pengalatdite/aplikasisederhana/MainActivityViewMvc.java
@@ -1,0 +1,11 @@
+package com.example.pengalatdite.aplikasisederhana;
+
+import android.view.View;
+
+import java.util.List;
+
+public interface MainActivityViewMvc {
+    View getRootView();
+
+    void setDataNews(List<DataNews> dataNews);
+}
