+++ app/src/main/java/com/example/pengalatdite/aplikasisederhana/ViewMvc.java
@@ -2,10 +2,8 @@ package com.example.pengalatdite.aplikasisederhana;
 
 import android.view.View;
 
-import java.util.List;
+public interface ViewMvc {
 
-public interface ViewMVC {
     View getRootView();
 
-    void setDataNews(List<DataNews> dataNews);
 }
